Refactor verify to be request ignorant.
The purpose is so a rack verification can take place outside the normal
rack request cycle. Create an instance of Rack::Recaptcha and call verify
with the right data and get back an answer.

My use case so I can do recaptcha verification in environments that don't
support Rack. I can get the info from the request myself and just give
Rack::Recaptcha the info I have extracted rather than having Rack do the
extraction.

Obviously running outside of the Rack environment is outside the scope of this
library (it is specifically for Rack) but making verify independent of the
request object seems like good design anyway (seperation of concerns) and then
we get the side benefit that I need.